+++ src/main/java/com/breakinblocks/nautec/compat/modonomicon/datagen/book/NautecGuide.java
@@ -3,6 +3,7 @@ package com.breakinblocks.nautec.compat.modonomicon.datagen.book;
 import com.klikli_dev.modonomicon.api.datagen.SingleBookSubProvider;
 import com.breakinblocks.nautec.Nautec;
 import com.breakinblocks.nautec.compat.modonomicon.datagen.book.nautec_guide.AquaticBiologyCategory;
+import com.breakinblocks.nautec.compat.modonomicon.datagen.book.nautec_guide.DeepEngineeringCategory;
 import com.breakinblocks.nautec.compat.modonomicon.datagen.book.nautec_guide.GettingStartedCategory;
 import com.breakinblocks.nautec.compat.modonomicon.datagen.book.nautec_guide.LaserAugmentationCategory;
 import com.breakinblocks.nautec.compat.modonomicon.datagen.book.nautec_guide.LaserChemistryCategory;
@@ -22,6 +23,7 @@ public class NautecGuide extends SingleBookSubProvider {
         add(new GettingStartedCategory(this).generate());
         add(new LaserChemistryCategory(this).generate());
         add(new AquaticBiologyCategory(this).generate());
+        add(new DeepEngineeringCategory(this).generate());
         add(new LaserAugmentationCategory(this).generate());
         add(new SubmarineCategory(this).generate());
     }